Remove remaining ViewControllerUI classes & Small UI Changes
