--- Heimdall/Protocols/ViewControllerUI.swift
@@ -1,13 +1,0 @@
-//
-//  ViewControllerUI.swift
-//  Heimdall
-//
-//  Created by Luis Reisewitz on 10.09.17.
-//  Copyright © 2017 Gnosis. All rights reserved.
-//
-
-import UIKit
-
-protocol ViewControllerUI {
-    var view: UIView { get }
-}
